refactor(dp-017a): split cli/verify.py into 3 focused modules
Extracted config construction and output handling from verify.py into dedicated
modules to reduce file size and improve maintainability:

- verify.py: 1079→492 lines (core CLI command logic)
- verify_config.py: NEW, 384 lines (config construction helpers)
- verify_output.py: NEW, 239 lines (output/display/progress helpers)

All linters (ruff) and type checkers (mypy) pass.
