+++ src/karenina/cli/verify.py
@@ -2,193 +2,33 @@
 Verify command for running benchmark verifications.
 
 This module implements the main 'karenina verify' command.
+Config building is in verify_config.py, output/display helpers in verify_output.py.
 """
 
 import time
-from datetime import datetime
 from pathlib import Path
 from typing import Annotated
 
 import typer
 from pydantic import ValidationError
 from rich.console import Console
-from rich.progress import (
-    BarColumn,
-    Progress,
-    SpinnerColumn,
-    TaskID,
-    TaskProgressColumn,
-    TextColumn,
-    TimeRemainingColumn,
-)
-from rich.prompt import Confirm, Prompt
 
-from karenina.adapters.manual import load_manual_traces_from_file
 from karenina.benchmark import Benchmark
-from karenina.benchmark.verification import (
-    export_verification_results_csv,
-    export_verification_results_json,
-)
-from karenina.benchmark.verification.batch_runner import generate_task_queue, run_verification_batch
-from karenina.schemas import FinishedTemplate, VerificationConfig, VerificationResult, VerificationResultSet
+from karenina.benchmark.verification.batch_runner import generate_task_queue
+from karenina.schemas import FinishedTemplate, VerificationConfig, VerificationResultSet
 from karenina.utils.progressive_save import ProgressiveSaveManager, TaskIdentifier, generate_task_manifest
 
-from .utils import (
-    cli_error,
-    create_export_job,
-    filter_templates_by_indices,
-    get_preset_path,
-    get_traces_path,
-    parse_question_indices,
-    validate_output_path,
+from .utils import cli_error, filter_templates_by_indices, parse_question_indices, validate_output_path
+from .verify_config import build_config_non_interactive
+from .verify_output import (
+    display_summary,
+    export_results,
+    prompt_for_output_interactive,
+    run_verification_with_progress,
+    validate_output_and_prompt,
 )
 
 console = Console()
-
-
-def _build_config_from_cli_args(
-    answering_model: str | None,
-    answering_provider: str | None,
-    answering_id: str,
-    parsing_model: str | None,
-    parsing_provider: str | None,
-    parsing_id: str,
-    temperature: float | None,
-    interface: str | None,
-    replicate_count: int | None,
-    abstention: bool,
-    sufficiency: bool,
-    embedding_check: bool,
-    deep_judgment: bool,
-    deep_judgment_rubric_mode: str,
-    deep_judgment_rubric_excerpts: bool,
-    deep_judgment_rubric_max_excerpts: int,
-    deep_judgment_rubric_fuzzy_threshold: float,
-    deep_judgment_rubric_retry_attempts: int,
-    deep_judgment_rubric_search: bool,
-    deep_judgment_rubric_search_tool: str,
-    deep_judgment_rubric_config: Path | None,
-    use_full_trace_for_template: bool,
-    use_full_trace_for_rubric: bool,
-    evaluation_mode: str,
-    embedding_threshold: float,
-    embedding_model: str,
-    async_execution: bool,
-    async_workers: int | None,
-    preset_config: VerificationConfig | None = None,
-    manual_traces_obj: object | None = None,
-) -> VerificationConfig:
-    """
-    Build VerificationConfig respecting hierarchy: CLI > preset > env > defaults.
-
-    Handles CLI-specific concerns (file loading for deep_judgment_rubric_config)
-    then delegates to VerificationConfig.from_overrides() for config construction.
-
-    Args:
-        CLI argument values (all optional)
-        preset_config: Optional preset configuration to use as base
-
-    Returns:
-        VerificationConfig with CLI overrides applied
-    """
-    # CLI-specific: load custom rubric config JSON if a file path was provided
-    rubric_config_dict = None
-    if deep_judgment_rubric_config is not None:
-        import json
-
-        try:
-            with open(deep_judgment_rubric_config) as f:
-                rubric_config_dict = json.load(f)
-        except (FileNotFoundError, json.JSONDecodeError, ValueError) as e:
-            raise ValueError(f"Failed to load custom rubric config from {deep_judgment_rubric_config}: {e}") from e
-
-    # Resolve interface for answering vs parsing:
-    # Parsing model should NOT use manual interface (only answering model can)
-    answering_interface = interface
-    parsing_interface = interface if interface != "manual" else "langchain"
-
-    return VerificationConfig.from_overrides(
-        base=preset_config,
-        # Model configuration
-        answering_model=answering_model,
-        answering_provider=answering_provider,
-        answering_id=answering_id,
-        answering_interface=answering_interface,
-        parsing_model=parsing_model,
-        parsing_provider=parsing_provider,
-        parsing_id=parsing_id,
-        parsing_interface=parsing_interface,
-        temperature=temperature,
-        manual_traces=manual_traces_obj,
-        # Execution settings
-        replicate_count=replicate_count,
-        # Feature flags
-        abstention=abstention,
-        sufficiency=sufficiency,
-        embedding_check=embedding_check,
-        deep_judgment=deep_judgment,
-        # Evaluation settings
-        evaluation_mode=evaluation_mode,
-        embedding_threshold=embedding_threshold,
-        embedding_model=embedding_model,
-        async_execution=async_execution,
-        async_workers=async_workers,
-        # Trace filtering
-        use_full_trace_for_template=use_full_trace_for_template,
-        use_full_trace_for_rubric=use_full_trace_for_rubric,
-        # Deep judgment rubric settings
-        deep_judgment_rubric_mode=deep_judgment_rubric_mode,
-        deep_judgment_rubric_excerpts=deep_judgment_rubric_excerpts,
-        deep_judgment_rubric_max_excerpts=deep_judgment_rubric_max_excerpts,
-        deep_judgment_rubric_fuzzy_threshold=deep_judgment_rubric_fuzzy_threshold,
-        deep_judgment_rubric_retry_attempts=deep_judgment_rubric_retry_attempts,
-        deep_judgment_rubric_search=deep_judgment_rubric_search,
-        deep_judgment_rubric_search_tool=deep_judgment_rubric_search_tool,
-        deep_judgment_rubric_config=rubric_config_dict,
-    )
-
-
-def _validate_output_and_prompt(
-    output: Path | None,
-    progressive_save: bool,
-    resume: bool,
-    interactive: bool,
-) -> tuple[Path | None, str | None]:
-    """
-    Validate output path and prompt for one if needed.
-
-    Returns:
-        Tuple of (output_path, output_format) - both may be None if no output wanted
-    """
-    output_format = None
-
-    # Validate output path upfront (fail fast)
-    if output:
-        try:
-            output_format = validate_output_path(output)
-        except ValueError as e:
-            cli_error(str(e), e)
-
-    # Validate progressive save requires output
-    if progressive_save and not output:
-        cli_error("--progressive-save requires --output to be specified")
-
-    # Prompt for output file if not specified (and not resuming or interactive)
-    if not output and not resume and not interactive:
-        suggested = f"verification_results_{datetime.now().strftime('%Y%m%d_%H%M%S')}.json"
-        console.print("[yellow]No output file specified.[/yellow]")
-        console.print(f"[dim]Suggested: {suggested}[/dim]")
-
-        user_input = typer.prompt("Output file path", default=suggested)
-        output = Path(user_input)
-
-        # Validate the prompted output path
-        try:
-            output_format = validate_output_path(output)
-        except ValueError as e:
-            cli_error(str(e), e)
-
-    return output, output_format
 
 
 def _handle_resume_mode(
@@ -215,7 +55,7 @@ def _handle_resume_mode(
         output_format = validate_output_path(output)
 
         console.print(
-            f"[green]✓ Resuming: {progressive_manager.completed_count}/{progressive_manager.total_tasks} "
+            f"[green]\u2713 Resuming: {progressive_manager.completed_count}/{progressive_manager.total_tasks} "
             f"tasks already completed[/green]"
         )
 
@@ -248,129 +88,9 @@ def _load_benchmark(benchmark_path: str) -> Benchmark:
     except (FileNotFoundError, ValidationError) as e:
         cli_error(f"loading benchmark: {e}", e)
 
-    console.print(f"[green]✓ Loaded benchmark: {benchmark.name or 'Unnamed'}[/green]")
+    console.print(f"[green]\u2713 Loaded benchmark: {benchmark.name or 'Unnamed'}[/green]")
     console.print(f"  Total questions: {len(benchmark.get_all_questions())}")
     return benchmark
-
-
-def _validate_cli_config_requirements(
-    interface: str | None,
-    answering_model: str | None,
-    parsing_model: str | None,
-    answering_provider: str | None,
-    parsing_provider: str | None,
-    manual_traces: Path | None,
-) -> list[str]:
-    """
-    Validate required CLI parameters when no preset is used.
-
-    Returns:
-        List of validation error messages (empty if valid)
-    """
-    validation_errors = []
-
-    if not interface:
-        validation_errors.append(
-            "--interface is required when not using a preset (langchain/openrouter/openai_endpoint)"
-        )
-
-    if not answering_model:
-        validation_errors.append("--answering-model is required when not using a preset")
-
-    if not parsing_model:
-        validation_errors.append("--parsing-model is required when not using a preset")
-
-    # Validate interface-specific requirements
-    if interface == "langchain":
-        if not answering_provider:
-            validation_errors.append(
-                "--answering-provider is required for langchain interface (e.g., openai, anthropic, google)"
-            )
-        if not parsing_provider:
-            validation_errors.append("--parsing-provider is required for langchain interface")
-    elif interface == "openai_endpoint":
-        console.print(
-            "[yellow]Note: OpenAI Endpoint interface requires endpoint_base_url. "
-            "Use interactive mode or a preset for this interface.[/yellow]"
-        )
-        validation_errors.append(
-            "OpenAI Endpoint interface is not supported via CLI args alone. Use --interactive or --preset."
-        )
-    elif interface == "manual":
-        if not manual_traces:
-            validation_errors.append(
-                "--manual-traces is required when --interface manual. "
-                "Provide a JSON file mapping question hashes to answer traces."
-            )
-
-    # Additional validation for manual traces
-    if manual_traces and not manual_traces.exists():
-        validation_errors.append(f"Manual traces file not found: {manual_traces}")
-
-    return validation_errors
-
-
-def _load_manual_traces(manual_traces: Path, benchmark: Benchmark) -> object:
-    """
-    Load manual traces from file.
-
-    Returns:
-        ManualTraces object
-
-    Raises:
-        typer.Exit: If loading fails
-    """
-    console.print(f"[cyan]Loading manual traces from {manual_traces}...[/cyan]")
-    try:
-        # Resolve trace file path
-        trace_file = get_traces_path(manual_traces)
-
-        # Load traces and create ManualTraces object
-        manual_traces_obj = load_manual_traces_from_file(trace_file, benchmark)
-
-        # Verify traces loaded and report to user
-        from karenina.adapters.manual import get_manual_trace_count
-
-        trace_count = get_manual_trace_count()
-        console.print(f"[green]✓ Loaded {trace_count} manual trace(s)[/green]")
-        return manual_traces_obj
-
-    except FileNotFoundError as e:
-        cli_error(str(e), e)
-    except Exception as e:
-        cli_error(f"loading manual traces: {e}", e)
-
-
-def _validate_deep_judgment_rubric_settings(
-    deep_judgment_rubric_mode: str,
-    deep_judgment_rubric_fuzzy_threshold: float,
-    deep_judgment_rubric_config: Path | None,
-) -> None:
-    """
-    Validate deep judgment rubric settings.
-
-    Raises:
-        typer.Exit: If validation fails
-    """
-    valid_modes = ["disabled", "enable_all", "use_checkpoint", "custom"]
-    if deep_judgment_rubric_mode not in valid_modes:
-        cli_error(
-            f"Invalid deep_judgment_rubric_mode '{deep_judgment_rubric_mode}'. Valid modes: {', '.join(valid_modes)}"
-        )
-
-    if not 0.0 <= deep_judgment_rubric_fuzzy_threshold <= 1.0:
-        cli_error("deep_judgment_rubric_fuzzy_threshold must be between 0.0 and 1.0")
-
-    if deep_judgment_rubric_mode == "use_checkpoint":
-        console.print(
-            "[yellow]Note: use_checkpoint mode requires deep judgment settings to be saved in the checkpoint file.[/yellow]"
-        )
-        console.print(
-            "[yellow]If the checkpoint doesn't have deep judgment config, traits will be disabled by default.[/yellow]"
-        )
-
-    if deep_judgment_rubric_mode == "custom" and not deep_judgment_rubric_config:
-        cli_error("custom mode requires --deep-judgment-rubric-config to be specified")
 
 
 def _filter_templates(
@@ -380,8 +100,6 @@ def _filter_templates(
 ) -> list[FinishedTemplate]:
     """
     Filter templates based on index-based criteria.
-
-    Note: Question ID filtering is handled upstream via Benchmark.get_finished_templates(question_ids=...).
 
     Returns:
         Filtered list of templates
@@ -411,41 +129,6 @@ def _filter_templates(
     return templates
 
 
-def _prompt_for_output_interactive(
-    progressive_save: bool,
-) -> tuple[Path | None, str | None, bool]:
-    """
-    Prompt for output file in interactive mode.
-
-    Returns:
-        Tuple of (output_path, output_format, progressive_save_updated)
-    """
-    if progressive_save:
-        console.print("\n[yellow]Progressive save requires an output file.[/yellow]")
-        save_to_file = True
-    else:
-        save_to_file = Confirm.ask("\nWould you like to save the verification results to a file?", default=True)
-
-    if save_to_file:
-        format_choice = Prompt.ask("Output format", choices=["json", "csv"], default="json")
-        default_filename = f"verification_results.{format_choice}"
-        output_filename = Prompt.ask("Output filename", default=default_filename)
-
-        output = Path(output_filename)
-
-        try:
-            output_format = validate_output_path(output)
-            console.print(f"[green]✓ Results will be saved to: {output}[/green]")
-            return output, output_format, progressive_save
-        except ValueError as e:
-            console.print(f"[yellow]Warning: {e}. Results will not be saved to file.[/yellow]")
-            if progressive_save:
-                console.print("[yellow]Disabling progressive save due to invalid output path.[/yellow]")
-            return None, None, False
-
-    return None, None, progressive_save
-
-
 def _initialize_progressive_save(
     output: Path,
     config: VerificationConfig,
@@ -471,277 +154,8 @@ def _initialize_progressive_save(
         output, config, benchmark_path, global_rubric=benchmark.get_global_rubric()
     )
     progressive_manager.initialize(task_manifest)
-    console.print(f"[green]✓ Progressive save enabled: {progressive_manager.tmp_path}[/green]")
+    console.print(f"[green]\u2713 Progressive save enabled: {progressive_manager.tmp_path}[/green]")
     return progressive_manager
-
-
-def _run_verification_with_progress(
-    templates: list[FinishedTemplate],
-    config: VerificationConfig,
-    benchmark: Benchmark,
-    progressive_manager: ProgressiveSaveManager | None,
-    show_progress: bool,
-) -> VerificationResultSet:
-    """
-    Run verification with optional progress bar display.
-
-    Returns:
-        VerificationResultSet containing all results
-    """
-    if show_progress:
-        total_verifications = (
-            len(templates) * len(config.answering_models) * len(config.parsing_models) * config.replicate_count
-        )
-
-        with Progress(
-            SpinnerColumn(),
-            TextColumn("[progress.description]{task.description}"),
-            BarColumn(),
-            TaskProgressColumn(),
-            TimeRemainingColumn(),
-            console=console,
-        ) as progress:
-            progress_task: TaskID = progress.add_task("Verifying questions...", total=total_verifications)
-
-            def progress_callback(current: int, _total: int, result: VerificationResult | None = None) -> None:
-                """Update progress bar on each verification completion."""
-                progress.update(progress_task, completed=current)
-                if result and result.verify_result is not None:
-                    status = "✓" if result.verify_result else "✗"
-                    progress.update(progress_task, description=f"Verifying questions... {status}")
-
-                if progressive_manager and result and result.metadata.timestamp:
-                    progressive_manager.add_result(result)
-
-            results = run_verification_batch(
-                templates=templates,
-                config=config,
-                run_name="cli-verification",
-                global_rubric=benchmark.get_global_rubric(),
-                progress_callback=progress_callback,
-            )
-    else:
-
-        def simple_progress_callback(_current: int, _total: int, result: VerificationResult | None = None) -> None:
-            """Save result incrementally (no progress bar display)."""
-            if progressive_manager and result and result.metadata.timestamp:
-                progressive_manager.add_result(result)
-
-        results = run_verification_batch(
-            templates=templates,
-            config=config,
-            run_name="cli-verification",
-            global_rubric=benchmark.get_global_rubric(),
-            progress_callback=simple_progress_callback if progressive_manager else None,
-        )
-
-    return results
-
-
-def _export_results(
-    output: Path,
-    output_format: str,
-    final_results: VerificationResultSet,
-    config: VerificationConfig,
-    benchmark: Benchmark,
-    start_time: float,
-    end_time: float,
-    progressive_manager: ProgressiveSaveManager | None,
-) -> None:
-    """
-    Export verification results to file.
-    """
-    console.print(f"\n[cyan]Exporting results to {output}...[/cyan]")
-
-    job = create_export_job(final_results, config, "cli-verification", start_time, end_time)
-
-    global_rubric = benchmark.get_global_rubric()
-    if output_format == "json":
-        export_data = export_verification_results_json(job, final_results, global_rubric)
-    else:  # csv
-        export_data = export_verification_results_csv(job, final_results, global_rubric)
-
-    output.write_text(export_data)
-    console.print(f"[green]✓ Results exported to {output}[/green]")
-
-    if progressive_manager:
-        progressive_manager.finalize()
-        console.print("[dim]Progressive save files cleaned up[/dim]")
-
-
-def _display_summary(
-    final_results: VerificationResultSet,
-    duration: float,
-) -> None:
-    """
-    Display verification summary.
-    """
-    total = len(final_results)
-    passed = sum(1 for r in final_results if r.template and r.template.verify_result)
-    execution_errors = sum(1 for r in final_results if not r.metadata.completed_without_errors)
-    failures = total - passed
-    not_passed = failures - execution_errors
-
-    console.print("\n[bold]Verification Summary:[/bold]")
-    console.print(f"  Total: {total}")
-    console.print(f"  Passed: [green]{passed}[/green]")
-    if failures > 0:
-        console.print(f"  Failures: [red]{failures}[/red]")
-        if execution_errors > 0:
-            console.print(f"    Execution Errors: [yellow]{execution_errors}[/yellow]")
-        if not_passed > 0:
-            console.print(f"    Not Passed: [red]{not_passed}[/red]")
-    else:
-        console.print("  Failures: [green]0[/green]")
-    console.print(f"  Duration: {duration:.2f}s")
-
-    if failures == 0:
-        console.print("\n[green]✓ Verification complete![/green]")
-    else:
-        console.print(f"\n[yellow]⚠ Verification complete with {failures} failure(s)[/yellow]")
-
-
-def _build_config_non_interactive(
-    preset: Path | None,
-    interface: str | None,
-    answering_model: str | None,
-    answering_provider: str | None,
-    answering_id: str,
-    parsing_model: str | None,
-    parsing_provider: str | None,
-    parsing_id: str,
-    temperature: float | None,
-    replicate_count: int | None,
-    abstention: bool,
-    sufficiency: bool,
-    embedding_check: bool,
-    deep_judgment: bool,
-    deep_judgment_rubric_mode: str,
-    deep_judgment_rubric_excerpts: bool,
-    deep_judgment_rubric_max_excerpts: int,
-    deep_judgment_rubric_fuzzy_threshold: float,
-    deep_judgment_rubric_retry_attempts: int,
-    deep_judgment_rubric_search: bool,
-    deep_judgment_rubric_search_tool: str,
-    deep_judgment_rubric_config: Path | None,
-    use_full_trace_for_template: bool,
-    use_full_trace_for_rubric: bool,
-    evaluation_mode: str,
-    embedding_threshold: float,
-    embedding_model: str,
-    async_execution: bool,
-    async_workers: int | None,
-    manual_traces: Path | None,
-    benchmark: Benchmark,
-    progressive_save: bool,
-) -> tuple[VerificationConfig, bool]:
-    """
-    Build verification config from CLI args and/or preset (non-interactive mode).
-
-    Returns:
-        Tuple of (config, progressive_save_updated)
-
-    Raises:
-        typer.Exit: If configuration errors occur
-    """
-    # Load preset if provided
-    preset_config = None
-    if preset:
-        console.print("[cyan]Loading preset...[/cyan]")
-        try:
-            preset_path = get_preset_path(str(preset))
-            preset_config = VerificationConfig.from_preset(preset_path)
-            console.print(f"[green]✓ Loaded preset from: {preset_path}[/green]")
-        except (FileNotFoundError, ValueError, ValidationError) as e:
-            cli_error(f"loading preset: {e}", e)
-
-    # Validate required parameters when no preset is used
-    if not preset_config:
-        validation_errors = _validate_cli_config_requirements(
-            interface, answering_model, parsing_model, answering_provider, parsing_provider, manual_traces
-        )
-        if validation_errors:
-            console.print("[red]Configuration errors:[/red]")
-            for error in validation_errors:
-                console.print(f"  [red]• {error}[/red]")
-            console.print("\n[dim]Run with --interactive for guided configuration[/dim]")
-            raise typer.Exit(code=1)
-
-    # Load manual traces if provided
-    manual_traces_obj = None
-    if manual_traces:
-        manual_traces_obj = _load_manual_traces(manual_traces, benchmark)
-
-    # Build config with CLI overrides
-    try:
-        config = _build_config_from_cli_args(
-            answering_model=answering_model,
-            answering_provider=answering_provider,
-            answering_id=answering_id,
-            parsing_model=parsing_model,
-            parsing_provider=parsing_provider,
-            parsing_id=parsing_id,
-            temperature=temperature,
-            interface=interface,
-            replicate_count=replicate_count,
-            abstention=abstention,
-            sufficiency=sufficiency,
-            embedding_check=embedding_check,
-            deep_judgment=deep_judgment,
-            deep_judgment_rubric_mode=deep_judgment_rubric_mode,
-            deep_judgment_rubric_excerpts=deep_judgment_rubric_excerpts,
-            deep_judgment_rubric_max_excerpts=deep_judgment_rubric_max_excerpts,
-            deep_judgment_rubric_fuzzy_threshold=deep_judgment_rubric_fuzzy_threshold,
-            deep_judgment_rubric_retry_attempts=deep_judgment_rubric_retry_attempts,
-            deep_judgment_rubric_search=deep_judgment_rubric_search,
-            deep_judgment_rubric_search_tool=deep_judgment_rubric_search_tool,
-            deep_judgment_rubric_config=deep_judgment_rubric_config,
-            use_full_trace_for_template=use_full_trace_for_template,
-            use_full_trace_for_rubric=use_full_trace_for_rubric,
-            evaluation_mode=evaluation_mode,
-            embedding_threshold=embedding_threshold,
-            embedding_model=embedding_model,
-            async_execution=async_execution,
-            async_workers=async_workers,
-            preset_config=preset_config,
-            manual_traces_obj=manual_traces_obj,
-        )
-
-        # Show what we're using
-        if preset_config:
-            has_cli_overrides = any(
-                [
-                    answering_model,
-                    answering_provider,
-                    parsing_model,
-                    parsing_provider,
-                    temperature,
-                    interface,
-                    replicate_count,
-                    abstention,
-                    embedding_check,
-                    deep_judgment,
-                    evaluation_mode,
-                    embedding_threshold,
-                    embedding_model,
-                    async_execution,
-                    async_workers,
-                ]
-            )
-            if has_cli_overrides:
-                console.print("[dim]CLI arguments will override preset values[/dim]")
-        else:
-            console.print("[dim]Building configuration from CLI arguments[/dim]")
-
-    except Exception as e:
-        cli_error(f"building configuration: {e}", e)
-
-    # Validate deep judgment rubric settings
-    _validate_deep_judgment_rubric_settings(
-        deep_judgment_rubric_mode, deep_judgment_rubric_fuzzy_threshold, deep_judgment_rubric_config
-    )
-
-    return config, progressive_save
 
 
 def _filter_templates_for_resume(
@@ -920,7 +334,7 @@ def verify(
         async_execution = not no_async
 
         # Step 1: Validate output path and prompt if needed
-        output, output_format = _validate_output_and_prompt(output, progressive_save, bool(resume), interactive)
+        output, output_format = validate_output_and_prompt(output, progressive_save, bool(resume), interactive)
 
         # Validate benchmark_path is provided unless resuming
         if not resume and not benchmark_path:
@@ -960,7 +374,7 @@ def verify(
             if not progressive_save:
                 progressive_save = progressive_save_interactive
         else:
-            config, progressive_save = _build_config_non_interactive(
+            config, progressive_save = build_config_non_interactive(
                 preset=preset,
                 interface=interface,
                 answering_model=answering_model,
@@ -1010,7 +424,7 @@ def verify(
 
         # Step 6: Prompt for output in interactive mode (if not already specified)
         if not output and interactive:
-            output, output_format, progressive_save = _prompt_for_output_interactive(progressive_save)
+            output, output_format, progressive_save = prompt_for_output_interactive(progressive_save)
 
         # Step 7: Initialize progressive save (if enabled and not resuming)
         if progressive_save and not resume:
@@ -1038,7 +452,7 @@ def verify(
         start_time = time.time()
         show_progress = show_progress_bar_interactive if show_progress_bar_interactive is not None else verbose
 
-        results = _run_verification_with_progress(templates, config, benchmark, progressive_manager, show_progress)
+        results = run_verification_with_progress(templates, config, benchmark, progressive_manager, show_progress)
 
         end_time = time.time()
         duration = end_time - start_time
@@ -1053,12 +467,12 @@ def verify(
 
         # Step 11: Export results (if output configured)
         if output and output_format:
-            _export_results(
+            export_results(
                 output, output_format, final_results, config, benchmark, start_time, end_time, progressive_manager
             )
 
         # Step 12: Display summary
-        _display_summary(final_results, duration)
+        display_summary(final_results, duration)
 
         raise typer.Exit(code=0)
